services/spider-workers/pyproject.toml
[project]
name = "estategap-spiders"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.62",
    "httpx>=0.27",
    "nats-py[nkeys]>=2.7",
    "parsel>=1.9",
    "playwright>=1.43",
    "playwright-stealth>=1.0.6",
    "prometheus-client>=0.20",
    "pydantic-settings>=2.3",
    "redis>=5.0",
    "structlog>=24",
    "estategap-common",
]

[dependency-groups]
dev = [
    "mypy",
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.35",
    "respx>=0.21",
    "ruff",
    "testcontainers[redis]>=4.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["estategap_spiders"]

[tool.uv.sources]
estategap-common = { path = "../../libs/common", editable = true }

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "ASYNC", "UP", "RUF"]

[tool.mypy]
python_version = "3.12"
strict = true
